requirements.txt
streamlit==1.52.0
openai==2.14.0
pillow>=10.0.0
streamlit-autorefresh>=1.0.1
streamlit-webrtc>=0.63.11
